requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0

# For better UI
pandas>=2.0.0

# Optional: For enhanced features
python-dotenv>=1.0.0
